+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). See [MAINTAINERS.md](./MAINTAINERS.md)
 for instructions to keep up to date.
 
+## Unreleased
+
+### Added
+
+* Added a `--latest-block-query` flag to the generic evm poller to allow for using either `finalized` or `latest` as parameter when querying for the latest block.
+
 ## v2.3.1
 
 ### Operators